production/quantum_notice_app/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0
google-generativeai==0.7.1
